requirements.txt
rdflib==7.0.0
networkx==3.2.1
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.4.0
scipy==1.12.0
joblib==1.3.2
pykeen==1.11.0
torch==2.4.1
qiskit==0.45.2
qiskit-machine-learning==0.7.1
qiskit-aer==0.13.3
qiskit-ibm-runtime==0.13.1
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
streamlit==1.32.0
matplotlib==3.8.3
seaborn==0.13.2
pyyaml==6.0.1
python-dotenv==1.0.1
jupyter==1.0.0